requirements.txt
peft==0.13.2
pandas==2.2.2
transformers==4.41.2
tqdm==4.66.5
scikit-learn==1.5.1
python-dotenv==1.1.0
